fuuck

+++ src/graphics/render_sector.c
@@ -6,7 +6,7 @@
 /*   By: mfischer <mfischer@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/11 00:24:25 by mfischer          #+#    #+#             */
-/*   Updated: 2019/05/11 00:29:54 by mfischer         ###   ########.fr       */
+/*   Updated: 2019/05/11 16:45:11 by mfischer         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ void		render_sector(t_sector *sector, t_libui_window *win, t_camera *cam)
 	while (++i < sector->objectnum)
 		render_object(cam, win, &sector->objects[i]);
 	i = -1;
-	while (++i < sector->surfacenum)
+	while (++i < (int)sector->surfacenum)
 		if (sector->surfaces[i].next_sector_id == -1)
 			render_polygons(sector->surfaces[i].polygons,
 							sector->surfaces[i].matrix, win, cam);
